+++ lib/src/shaders/dither_shader_wrap.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_shaders/flutter_shaders.dart';
 import 'package:material_palette/src/shader_animation.dart';
 import 'package:material_palette/src/shader_definitions.dart';
 import 'package:material_palette/src/shader_params.dart';
@@ -25,6 +26,9 @@ class DitherShaderWrap extends StatelessWidget {
   final double time;
   final ShaderAnimationConfig? animationConfig;
   final bool cache;
+
+  static Future<void> precacheShader() => ShaderBuilder.precacheShader(
+      'packages/material_palette/shaders/dither_wrap.frag');
 
   @override
   Widget build(BuildContext context) {